+++ app/helpers/app_configs_helper.rb
@@ -2,18 +2,10 @@
 
 module AppConfigsHelper
   def current_favicon_and_options
-    return unless @app_configs.favicon.attached?
-    '<br><br><strong>Current Favicon: </strong> '\
-        "#{current_favicon}"
-  end
-
-  private
-
-  def current_favicon
-    resized_favicon = @app_configs
-                      .favicon
-                      .variant(resize('150x150'))
-
-    image_tag resized_favicon
+    if @app_configs.favicon.attached?
+      @app_configs.favicon.variant(resize: '75x75')
+    else
+      'favicon.ico?v=2'
+    end
   end
 end